updated HTTP classes and put hello.js in a dedicated module

+++ public/packages/resiway/actions/user/signup.php
@@ -8,6 +8,7 @@ use easyobject\orm\PersistentDataManager as PersistentDataManager;
 use html\HtmlTemplate as HtmlTemplate;
 use maxmind\geoip\GeoIP as GeoIP;
 
+
 // force silent mode (debug output would corrupt json data)
 set_silent(true);
 
@@ -16,20 +17,40 @@ $params = QNLib::announce(
 	array(	
     'description'	=>	"Attempt to register a new user.",
     'params' 		=>	array(
-                        'login'	    =>  array(
-                                        'description'   => 'email address of the user.',
-                                        'type'          => 'string', 
-                                        'required'      => true
+                        'login'	        =>  array(
+                                            'description'   => 'email address of the user.',
+                                            'type'          => 'string', 
+                                            'required'      => true
                                         ),
-                        'firstname'	=>  array(
-                                        'description'   => 'user\'s firstname',
-                                        'type'          => 'string', 
-                                        'required'      => true
+                        'firstname'	    =>  array(
+                                            'description'   => 'user\'s firstname',
+                                            'type'          => 'string', 
+                                            'required'      => true
                                         ),
-                        'lang'	    =>  array(
-                                        'description'   => 'user\'s prefered language',
-                                        'type'          => 'string', 
-                                        'default'       => 'fr'
+                        'lastname'	    =>  array(
+                                            'description'   => 'user\'s lastname',
+                                            'type'          => 'string', 
+                                            'default'       => ''
+                                        ),                                        
+                        'lang'	        =>  array(
+                                            'description'   => 'user\'s prefered language',
+                                            'type'          => 'string', 
+                                            'default'       => 'fr'
+                                        ),
+                        'account_type'  =>  array(
+                                            'description'   => 'origin of user account.',
+                                            'type'          => 'string', 
+                                            'default'       => 'resiway'
+                                        ),
+                        'avatar_url'    =>  array(
+                                            'description'   => 'URL of the user avatar.',
+                                            'type'          => 'string', 
+                                            'default'       => ''
+                                        ),
+                        'send_confirm'  =>  array(
+                                            'description'   => 'Flag telling if we need to send a confirmation email.',
+                                            'type'          => 'boolean', 
+                                            'default'       => true
                                         )                                        
                         )
 	)
@@ -38,16 +59,20 @@ $params = QNLib::announce(
 
 list($result, $error_message_ids) = [true, []];
 
-list($action_name, $login, $firstname, $language) = [ 
+list($action_name, $login, $firstname, $lastname, $language, $account_type, $avatar_url, $send_confirm) = [ 
     'resiway_user_signup',
     strtolower(trim($params['login'])),
     $params['firstname'],
-    $params['lang']
+    $params['lastname'],
+    $params['lang'],
+    $params['account_type'],
+    $params['avatar_url'],
+    $params['send_confirm']    
 ];
 
 $messages_folder = '../spool';
 
-try {
+try {    
     $om = &ObjectManager::getInstance();
     $pdm = &PersistentDataManager::getInstance();
 
@@ -70,18 +95,22 @@ try {
         $password .= sprintf("%x", rand(0, 15)) ;
     }
     
-    // generate avatar URL using identicon with a random hash
-    $avatar_url = 'https://www.gravatar.com/avatar/'.md5($firstname.rand()).'?d=identicon&s=@size';
+    if(strlen($avatar_url) <= 0) {
+        // generate avatar URL using identicon with a random hash
+        $avatar_url = 'https://www.gravatar.com/avatar/'.md5($firstname.rand()).'?d=identicon&s=@size';
+    }
     // get requesting IP geo location
     $location = GeoIP::getLocationFromIP($_SERVER['REMOTE_ADDR']);
     // init creation array with new user info
     $user_data = [
                     'login'         => $login, 
                     'password'      => $password, 
-                    'firstname'     => $firstname, 
+                    'firstname'     => $firstname,
+                    'lastname'      => $lastname,
                     'language'      => $language, 
                     'avatar_url'    => $avatar_url,                    
-                    'location'      => $location->city
+                    'location'      => $location->city,
+                    'account_type'  => $account_type
                    ];
     // assign returned country code only if consistent              
     if( strlen($location->country_code) == 2) $user_data['country'] = strtoupper($location->country_code);
@@ -100,47 +129,49 @@ try {
     // update session data
     $pdm->set('user_id', $user_id);
     
-    // retrieve newly created user
-    $user_data = ResiAPI::loadUserPrivate($user_id);
-    // we need the password to generate confirmation code in the email
-    $user_data['password'] = $password;
+    if($send_confirm) {
+        // retrieve newly created user
+        $user_data = ResiAPI::loadUserPrivate($user_id);
+        // we need the password to generate confirmation code in the email
+        $user_data['password'] = $password;
 
-    
-    // subject of the email should be defined in the template, as a <var> tag holding a 'title' attribute
-    $subject = '';
-    // read template according to user prefered language
-    $file = "packages/resiway/i18n/{$user_data['language']}/mail_user_confirm.html";
-    if(!($html = @file_get_contents($file, FILE_TEXT))) throw new Exception("action_failed", QN_ERROR_UNKNOWN);
-    $template = new HtmlTemplate($html, [
-                                'subject'		=>	function ($params, $attributes) use (&$subject) {
-                                                        $subject = $attributes['title'];
-                                                        return '';
-                                                    },
-                                'username'		=>	function ($params, $attributes) {
-                                                        return $params['firstname'];
-                                                    },
-                                'confirm_url'	=>	function ($params, $attributes) {
-                                                        $code = ResiAPI::credentialsEncode($params['login'],$params['password']);
-                                                        $url = QNlib::get_url(true, false)."user/confirm/{$code}";
-                                                        return "<a href=\"$url\">{$attributes['title']}</a>";
-                                                    }
-                                ], 
-                                $user_data);
-    // parse template as html
-    $body = $template->getHtml();
-    
-    /**
-    * message files format is: 11 digits (user unique identifier) with 3 digits extension in case of multiple files
-    */
-    $temp = sprintf("%011d", $user_id);
-    $filename = $temp;
-    $i = 0;
-    while(file_exists("$messages_folder/{$filename}")) {
-        $filename = sprintf("%s.%03d", $temp, ++$i);
+        
+        // subject of the email should be defined in the template, as a <var> tag holding a 'title' attribute
+        $subject = '';
+        // read template according to user prefered language
+        $file = "packages/resiway/i18n/{$user_data['language']}/mail_user_confirm.html";
+        if(!($html = @file_get_contents($file, FILE_TEXT))) throw new Exception("action_failed", QN_ERROR_UNKNOWN);
+        $template = new HtmlTemplate($html, [
+                                    'subject'		=>	function ($params, $attributes) use (&$subject) {
+                                                            $subject = $attributes['title'];
+                                                            return '';
+                                                        },
+                                    'username'		=>	function ($params, $attributes) {
+                                                            return $params['firstname'];
+                                                        },
+                                    'confirm_url'	=>	function ($params, $attributes) {
+                                                            $code = ResiAPI::credentialsEncode($params['login'],$params['password']);
+                                                            $url = QNlib::get_url(true, false)."user/confirm/{$code}";
+                                                            return "<a href=\"$url\">{$attributes['title']}</a>";
+                                                        }
+                                    ], 
+                                    $user_data);
+        // parse template as html
+        $body = $template->getHtml();
+        
+        /**
+        * message files format is: 11 digits (user unique identifier) with 3 digits extension in case of multiple files
+        */
+        $temp = sprintf("%011d", $user_id);
+        $filename = $temp;
+        $i = 0;
+        while(file_exists("$messages_folder/{$filename}")) {
+            $filename = sprintf("%s.%03d", $temp, ++$i);
+        }
+        // data consists of parsed template and subject
+        $json = json_encode(array("subject" => $subject, "body" => $body), JSON_PRETTY_PRINT);
+        file_put_contents("$messages_folder/$filename", $json);
     }
-    // data consists of parsed template and subject
-    $json = json_encode(array("subject" => $subject, "body" => $body), JSON_PRETTY_PRINT);
-    file_put_contents("$messages_folder/$filename", $json);
     
     // log user registration
     ResiAPI::registerAction($user_id, $action_name, 'resiway\User', $user_id);
